@@ -1,6 +1,6 @@
 {
   "name": "anekdotes/menus",
-  "version": "1.2.0",
+  "version": "2.0.0",
   "type": "library",
   "description": "Utility to generate a static array full of links",
   "keywords": [
@@ -18,12 +18,11 @@
     }
   ],
   "require": {
-    "php": ">=5.6.0",
-    "anekdotes/support": "<2.0.0"
+    "php": ">=7.4.0",
+    "anekdotes/support": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "~5.4",
-    "codacy/coverage": "^1.0"
+    "phpunit/phpunit": "^9"
   },
   "autoload": {
       "classmap": [
